+++ demos/_LHC_acc.cpp
@@ -19,7 +19,7 @@ using namespace std;
 
 // Simulation parameters
 // --------------------------------------------------------
-const long N_b = (long) 1.2e9;                // Intensity
+const long long N_b = 1.2e9;                // Intensity
 int N_p = 600000;                           // Macro-particles
 
 // Machine and RF parameters
